// src/api/generated/models/landmarkDataStructure.ts
/**
 * Generated by orval v7.18.0 🍺
 * Do not edit manually.
 * History Trivia Backend API
 * API for History Trivia Chronokwiz game
 * OpenAPI spec version: 1.0.0
 */
import type { LandmarkDataStructureTypeOfPointOfOccurence } from './landmarkDataStructureTypeOfPointOfOccurence';

export interface LandmarkDataStructure {
  id?: number;
  hintDescription?: string;
  fullDescription?: string;
  pointOfOccurence?: string;
  internetLink?: string;
  category?: string;
  isMoved?: boolean;
  isConfirmed?: boolean;
  typeOfPointOfOccurence?: LandmarkDataStructureTypeOfPointOfOccurence;
}